fix context switch

+++ arch/risc-v/src/common/up_internal.h
@@ -138,7 +138,7 @@ void up_allocat_eheap(FAR void **heap_start, size_t *heap_size);
 /* IRQ initialization *******************************************************/
 
 void up_irqinitialize(void);
-void up_copystate(uint32_t *dest, uint32_t *src);
+void up_copystate(uint64_t *dest, uint64_t *src);
 void up_dumpstate(void);
 void up_sigdeliver(void);
 int up_swint(int irq, FAR void *context, FAR void *arg);
